requirements.txt
#TODO: Find out which of these are actually needed
#colorzero==1.1
#gpiozero==1.5.1
#pkg-resources==0.0.0
#RPi.GPIO==0.7.0
#smbus2==0.3.0
pynmea2==1.15.0
pyserial==3.4
